feat(hw/display): added linux support
